feat: port hardened stealth script into a shared module (#62)
Ports greedysearch-pi's validated injectHeadlessStealth (Sannysoft 20/20
clean, identical CreepJS fingerprints headless vs visible) into
extractors/stealth-script.mjs, a single source of truth now consumed by
both extractors/common.mjs (CDP) and src/fetch.ts's Playwright fallback,
eliminating the drift between two independently-maintained copies.

src/fetch.ts loads the shared .mjs script via a dynamic import resolved
at runtime (walking up from the running module to find
extractors/stealth-script.mjs), since tsc mirrors src/ one directory
level deeper into dist/src/ than the source tree, and extractors/ is
shipped as-is alongside dist/ rather than compiled.
